backend/requirements.txt
django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
fal-client==0.13.2
whitenoise==6.12.0
gunicorn==25.3.0
Pillow==12.1.1
